+++ ida_analyze_util.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 """Shared utility helpers for IDA analyze scripts."""
 
-import asyncio
 import json
 import math
 import ntpath
@@ -27,6 +26,25 @@ except Exception:
     call_llm_text = None
     normalize_optional_effort = None
     normalize_optional_temperature = None
+
+import ida_llm_decompile as _ida_llm_decompile
+from ida_llm_decompile import (
+    _build_llm_decompile_request_cache_key,
+    _debug_print_json,
+    _debug_print_multiline,
+    _derive_module_name,
+    _empty_llm_decompile_result,
+    _extract_llm_error_status_code,
+    _is_transient_llm_error,
+    _normalize_llm_entries,
+    _normalize_llm_retry_attempts,
+    _normalize_llm_retry_delay,
+    _normalize_llm_struct_offset_entries,
+    _parse_yaml_mapping,
+    _render_llm_decompile_blocks,
+    _resolve_llm_decompile_template_value,
+    parse_llm_decompile_response,
+)
 
 
 def _is_remote_absolute_path(path):
@@ -1005,128 +1023,8 @@ def write_struct_offset_yaml(path, data):
         )
 
 
-def _empty_llm_decompile_result():
-    return {
-        "found_vcall": [],
-        "found_call": [],
-        "found_funcptr": [],
-        "found_gv": [],
-        "found_struct_offset": [],
-    }
-
-
-def _normalize_llm_retry_attempts(value, default=3):
-    try:
-        attempts = int(value)
-    except (TypeError, ValueError):
-        attempts = int(default)
-    return max(1, attempts)
-
-
-def _normalize_llm_retry_delay(value, default, minimum=0.0):
-    try:
-        delay = float(value)
-    except (TypeError, ValueError):
-        delay = float(default)
-    if delay < minimum:
-        return minimum
-    return delay
-
-
-def _extract_llm_error_status_code(exc):
-    for source in (exc, getattr(exc, "response", None)):
-        if source is None:
-            continue
-        status_code = getattr(source, "status_code", None)
-        if status_code is None:
-            continue
-        try:
-            return int(status_code)
-        except (TypeError, ValueError):
-            continue
-    return None
-
-
-def _is_transient_llm_error(exc):
-    status_code = _extract_llm_error_status_code(exc)
-    if status_code == 429 or (
-        status_code is not None and 500 <= status_code < 600
-    ):
-        return True
-
-    message = str(exc or "").lower()
-    retryable_fragments = (
-        "transport received error",
-        "timeout",
-        "timed out",
-        "read timeout",
-        "rate limit",
-        "rate_limit",
-        "too many requests",
-        "http 429",
-        "status 429",
-        "status_code=429",
-        " 429",
-        "http 500",
-        "http 502",
-        "http 503",
-        "http 504",
-        "status 500",
-        "status 502",
-        "status 503",
-        "status 504",
-        "status_code=500",
-        "status_code=502",
-        "status_code=503",
-        "status_code=504",
-        " 500",
-        " 502",
-        " 503",
-        " 504",
-        "server error",
-        "service unavailable",
-        "temporarily unavailable",
-    )
-    return any(fragment in message for fragment in retryable_fragments)
-
-
 def _get_preprocessor_scripts_dir():
     return Path(__file__).resolve().parent / "ida_preprocessor_scripts"
-
-
-def _derive_module_name(new_binary_dir):
-    if not new_binary_dir:
-        return ""
-    return os.path.basename(os.path.normpath(str(new_binary_dir)))
-
-
-def _resolve_llm_decompile_template_value(value, platform, module_name=None):
-    resolved = str(value or "")
-    platform_text = str(platform or "").strip()
-    if platform_text:
-        resolved = resolved.replace("{platform}", platform_text)
-    module_text = str(module_name or "").strip()
-    if module_text:
-        resolved = resolved.replace("{module_name}", module_text)
-    return resolved
-
-
-def _debug_print_multiline(label, text, debug=False):
-    if not debug:
-        return
-    print(f"    Preprocess: {label} BEGIN")
-    print(str(text or "<empty>"))
-    print(f"    Preprocess: {label} END")
-
-
-def _debug_print_json(label, value, debug=False):
-    if not debug:
-        return
-    try:
-        rendered = json.dumps(value, indent=2, ensure_ascii=False, sort_keys=False)
-    except Exception:
-        rendered = repr(value)
-    _debug_print_multiline(label, rendered, debug=True)
 
 
 def _debug_format_addr_preview(values, limit=4):
@@ -1237,70 +1135,6 @@ def _build_llm_decompile_specs_map(llm_decompile_specs, debug=False):
         specs_map[func_name] = [llm_spec]
 
     return specs_map
-
-
-def _parse_yaml_mapping(text):
-    if yaml is None:
-        return None
-    try:
-        parsed = yaml.load(text, Loader=yaml.BaseLoader)
-    except yaml.YAMLError:
-        return None
-    if parsed is None:
-        return {}
-    if isinstance(parsed, dict):
-        return parsed
-    return None
-
-
-def _normalize_llm_entries(entries, required_keys):
-    if isinstance(entries, (str, bytes, bytearray)) or not isinstance(entries, (list, tuple)):
-        return []
-
-    normalized = []
-    for entry in entries:
-        if not isinstance(entry, dict):
-            continue
-        item = {}
-        valid = True
-        for key in required_keys:
-            value = str(entry.get(key, "")).strip()
-            if not value:
-                valid = False
-                break
-            item[key] = value
-        if valid:
-            normalized.append(item)
-    return normalized
-
-
-def _normalize_llm_struct_offset_entries(entries):
-    if isinstance(entries, (str, bytes, bytearray)) or not isinstance(entries, (list, tuple)):
-        return []
-
-    normalized = []
-    for entry in entries:
-        if not isinstance(entry, dict):
-            continue
-        item = {}
-        valid = True
-        for key in ("insn_va", "insn_disasm", "offset", "struct_name", "member_name"):
-            value = str(entry.get(key, "")).strip()
-            if not value:
-                valid = False
-                break
-            item[key] = value
-        if not valid:
-            continue
-
-        size_value = entry.get("size")
-        if size_value is not None:
-            size_text = str(size_value).strip()
-            if size_text:
-                item["size"] = size_text
-
-        normalized.append(item)
-    return normalized
 
 
 def _extract_slot_only_vfunc_candidates_from_llm_result(
@@ -2387,89 +2221,6 @@ async def _load_llm_decompile_target_details_via_mcp(
     return target_items
 
 
-def _render_llm_decompile_blocks(reference_items, target_items):
-    def _normalize_items(items):
-        if isinstance(items, dict):
-            return [items]
-        if isinstance(items, (tuple, list)):
-            return [item for item in items if isinstance(item, dict)]
-        return []
-
-    def _render_block(block_kind, item):
-        func_name = str(item.get("func_name", "") or "").strip() or "<unknown>"
-        disasm_code = str(item.get("disasm_code", "") or "")
-        procedure = str(item.get("procedure", "") or "")
-        return (
-            f"### {block_kind} Function: {func_name}\n\n"
-            "**Disassembly**\n\n"
-            f"```c\n{disasm_code}\n```\n\n"
-            "**Procedure**\n\n"
-            f"```c\n{procedure}\n```"
-        )
-
-    reference_blocks = "\n\n".join(
-        _render_block("Reference", item)
-        for item in _normalize_items(reference_items)
-    )
-    target_blocks = "\n\n".join(
-        _render_block("Target", item)
-        for item in _normalize_items(target_items)
-    )
-    return reference_blocks, target_blocks
-
-
-def parse_llm_decompile_response(response_text):
-    response_text = str(response_text or "").strip()
-    if not response_text:
-        return _empty_llm_decompile_result()
-
-    candidates = []
-    for match in re.finditer(
-        r"```(?:yaml|yml)[ \t]*\n?(.*?)```",
-        response_text,
-        re.IGNORECASE | re.DOTALL,
-    ):
-        candidates.append(match.group(1).strip())
-    if not candidates:
-        for match in re.finditer(r"```[ \t]*\n(.*?)```", response_text, re.DOTALL):
-            candidates.append(match.group(1).strip())
-    if not candidates:
-        candidates.append(response_text)
-
-    parsed = None
-    for candidate in candidates:
-        if not candidate:
-            continue
-        parsed = _parse_yaml_mapping(candidate)
-        if parsed is not None:
-            break
-
-    if not isinstance(parsed, dict):
-        return _empty_llm_decompile_result()
-
-    return {
-        "found_vcall": _normalize_llm_entries(
-            parsed.get("found_vcall", []),
-            ("insn_va", "insn_disasm", "vfunc_offset", "func_name"),
-        ),
-        "found_call": _normalize_llm_entries(
-            parsed.get("found_call", []),
-            ("insn_va", "insn_disasm", "func_name"),
-        ),
-        "found_funcptr": _normalize_llm_entries(
-            parsed.get("found_funcptr", []),
-            ("insn_va", "insn_disasm", "funcptr_name"),
-        ),
-        "found_gv": _normalize_llm_entries(
-            parsed.get("found_gv", []),
-            ("insn_va", "insn_disasm", "gv_name"),
-        ),
-        "found_struct_offset": _normalize_llm_struct_offset_entries(
-            parsed.get("found_struct_offset", []),
-        ),
-    }
-
-
 def _prepare_llm_decompile_request(
     func_name,
     llm_decompile_specs_map,
@@ -2478,279 +2229,17 @@ def _prepare_llm_decompile_request(
     new_binary_dir=None,
     debug=False,
 ):
-    module_name = _derive_module_name(new_binary_dir)
-    llm_spec = (llm_decompile_specs_map or {}).get(func_name)
-    if llm_spec is None:
-        return None
-
-    if yaml is None:
-        if debug:
-            print("    Preprocess: PyYAML is required for llm_decompile fallback")
-        return None
-
-    if not isinstance(llm_config, dict):
-        if debug:
-            print(f"    Preprocess: llm_config missing or invalid for {func_name}")
-        return None
-
-    model = str(llm_config.get("model", "")).strip()
-    if not model:
-        if debug:
-            print(f"    Preprocess: llm_config.model missing for {func_name}")
-        return None
-
-    api_key = llm_config.get("api_key")
-    base_url = llm_config.get("base_url")
-    fake_as = str(llm_config.get("fake_as") or "").strip().lower() or None
-
-    temperature = llm_config.get("temperature")
-    if temperature is not None:
-        if callable(normalize_optional_temperature):
-            try:
-                temperature = normalize_optional_temperature(
-                    temperature,
-                    "llm_config.temperature",
-                )
-            except ValueError as exc:
-                if debug:
-                    print(
-                        f"    Preprocess: invalid llm_decompile temperature for "
-                        f"{func_name}: {exc}"
-                    )
-                return None
-        else:
-            try:
-                temperature = float(temperature)
-            except (TypeError, ValueError):
-                if debug:
-                    print(
-                        f"    Preprocess: invalid llm_decompile temperature for "
-                        f"{func_name}: {temperature!r}"
-                    )
-                return None
-
-    if callable(normalize_optional_effort):
-        try:
-            effort = normalize_optional_effort(
-                llm_config.get("effort"),
-                "llm_config.effort",
-            )
-        except ValueError as exc:
-            if debug:
-                print(
-                    f"    Preprocess: invalid llm_decompile effort for "
-                    f"{func_name}: {exc}"
-                )
-            return None
-    else:
-        effort = str(llm_config.get("effort") or "").strip().lower() or "medium"
-
-    max_retries = _normalize_llm_retry_attempts(
-        llm_config.get("max_retries"),
-        default=3,
+    return _ida_llm_decompile._prepare_llm_decompile_request(
+        func_name,
+        llm_decompile_specs_map,
+        llm_config,
+        platform=platform,
+        new_binary_dir=new_binary_dir,
+        debug=debug,
+        preprocessor_scripts_dir=_get_preprocessor_scripts_dir(),
+        normalize_temperature_func=normalize_optional_temperature,
+        normalize_effort_func=normalize_optional_effort,
     )
-    retry_initial_delay = _normalize_llm_retry_delay(
-        llm_config.get("retry_initial_delay"),
-        default=1.0,
-    )
-    retry_backoff_factor = _normalize_llm_retry_delay(
-        llm_config.get("retry_backoff_factor"),
-        default=2.0,
-        minimum=1.0,
-    )
-    retry_max_delay = _normalize_llm_retry_delay(
-        llm_config.get("retry_max_delay"),
-        default=8.0,
-    )
-
-    if isinstance(llm_spec, dict):
-        llm_specs = [llm_spec]
-    elif isinstance(llm_spec, (tuple, list)) and llm_spec:
-        llm_specs = list(llm_spec)
-    else:
-        if debug:
-            print(f"    Preprocess: invalid llm_decompile spec for {func_name}")
-        return None
-
-    if not all(isinstance(spec, dict) for spec in llm_specs):
-        if debug:
-            print(f"    Preprocess: invalid llm_decompile spec for {func_name}")
-        return None
-
-    prompt_value = llm_specs[0].get("prompt_path")
-    if not isinstance(prompt_value, str) or not prompt_value:
-        if debug:
-            print(
-                "    Preprocess: invalid llm_decompile prompt path for "
-                f"{func_name}: {prompt_value!r}"
-            )
-        return None
-
-    for current_spec in llm_specs[1:]:
-        current_prompt_value = current_spec.get("prompt_path")
-        if current_prompt_value != prompt_value:
-            if debug:
-                print(
-                    "    Preprocess: mixed llm_decompile prompt paths for "
-                    f"{func_name}: {prompt_value!r} != {current_prompt_value!r}"
-                )
-            return None
-
-    scripts_dir = _get_preprocessor_scripts_dir()
-    prompt_path = Path(
-        _resolve_llm_decompile_template_value(
-            prompt_value,
-            platform,
-            module_name=module_name,
-        )
-    )
-    if not prompt_path.is_absolute():
-        prompt_path = scripts_dir / prompt_path
-    prompt_path = _absolute_path_preserve_spelling(prompt_path)
-
-    if not prompt_path.is_file():
-        if debug:
-            print(
-                f"    Preprocess: llm_decompile prompt missing for {func_name}: "
-                f"{prompt_path}"
-            )
-        return None
-
-    try:
-        prompt_template = prompt_path.read_text(encoding="utf-8")
-    except OSError as exc:
-        if debug:
-            print(
-                f"    Preprocess: failed to read llm_decompile prompt for "
-                f"{func_name}: {exc}"
-            )
-        return None
-
-    reference_items = []
-    reference_yaml_paths = []
-    target_func_names = []
-    for current_spec in llm_specs:
-        reference_value = current_spec.get("reference_yaml_path")
-        if not isinstance(reference_value, str) or not reference_value:
-            if debug:
-                print(
-                    "    Preprocess: invalid llm_decompile reference path for "
-                    f"{func_name}: {reference_value!r}"
-                )
-            return None
-        reference_yaml_path = Path(
-            _resolve_llm_decompile_template_value(
-                reference_value,
-                platform,
-                module_name=module_name,
-            )
-        )
-        if not reference_yaml_path.is_absolute():
-            reference_yaml_path = scripts_dir / reference_yaml_path
-        reference_yaml_path = _absolute_path_preserve_spelling(reference_yaml_path)
-
-        if not reference_yaml_path.is_file():
-            if debug:
-                print(
-                    f"    Preprocess: llm_decompile reference missing for "
-                    f"{func_name}: {reference_yaml_path}"
-                )
-            return None
-
-        try:
-            with open(reference_yaml_path, "r", encoding="utf-8") as handle:
-                reference_data = yaml.safe_load(handle) or {}
-        except Exception as exc:
-            if debug:
-                print(
-                    f"    Preprocess: failed to read llm_decompile reference for "
-                    f"{func_name}: {exc}"
-                )
-            return None
-
-        if not isinstance(reference_data, dict):
-            if debug:
-                print(
-                    f"    Preprocess: invalid llm_decompile reference payload for "
-                    f"{func_name}"
-                )
-            return None
-
-        target_func_name = str(reference_data.get("func_name", "") or "").strip()
-        if not target_func_name:
-            if debug:
-                print(
-                    "    Preprocess: llm_decompile reference func_name missing for "
-                    f"{func_name}"
-                )
-            return None
-
-        reference_items.append(reference_data)
-        reference_yaml_paths.append(os.fspath(reference_yaml_path))
-        target_func_names.append(target_func_name)
-
-    reference_data = reference_items[0]
-    reference_yaml_path = reference_yaml_paths[0]
-    target_func_name = target_func_names[0]
-
-    if debug:
-        print(
-            f"    Preprocess: llm_decompile request ready for {func_name}: "
-            f"platform={str(platform or '').strip() or '<empty>'}, "
-            f"model={model}, prompt_path={prompt_path}, "
-            f"reference_yaml_paths={reference_yaml_paths}"
-        )
-        _debug_print_json(
-            f"llm_decompile reference payloads for {func_name}",
-            reference_items,
-            debug=True,
-        )
-
-    return {
-        "model": model,
-        "prompt_path": os.fspath(prompt_path),
-        "reference_items": reference_items,
-        "reference_yaml_paths": reference_yaml_paths,
-        "target_func_names": target_func_names,
-        "reference_yaml_path": reference_yaml_path,
-        "prompt_template": prompt_template,
-        "target_func_name": target_func_name,
-        "disasm_for_reference": str(reference_data.get("disasm_code", "") or ""),
-        "procedure_for_reference": str(reference_data.get("procedure", "") or ""),
-        "temperature": temperature,
-        "effort": effort,
-        "api_key": api_key,
-        "base_url": base_url,
-        "fake_as": fake_as,
-        "max_retries": max_retries,
-        "retry_initial_delay": retry_initial_delay,
-        "retry_backoff_factor": retry_backoff_factor,
-        "retry_max_delay": retry_max_delay,
-    }
-
-
-def _build_llm_decompile_request_cache_key(llm_request):
-    if not isinstance(llm_request, dict):
-        return None
-    model = str(llm_request.get("model", "")).strip()
-    prompt_path = str(llm_request.get("prompt_path", "")).strip()
-    reference_yaml_paths = llm_request.get("reference_yaml_paths")
-    if reference_yaml_paths is None:
-        reference_yaml_path = str(llm_request.get("reference_yaml_path", "")).strip()
-        reference_yaml_paths = [reference_yaml_path] if reference_yaml_path else []
-    if isinstance(reference_yaml_paths, str):
-        reference_yaml_paths = [reference_yaml_paths]
-    if not isinstance(reference_yaml_paths, (tuple, list)):
-        return None
-    reference_yaml_paths = tuple(
-        str(reference_yaml_path).strip()
-        for reference_yaml_path in reference_yaml_paths
-        if str(reference_yaml_path).strip()
-    )
-    if not model or not prompt_path or not reference_yaml_paths:
-        return None
-    temperature = llm_request.get("temperature")
-    return model, prompt_path, reference_yaml_paths, temperature
 
 
 async def call_llm_decompile(
@@ -2777,172 +2266,32 @@ async def call_llm_decompile(
     retry_max_delay=None,
     debug=False,
 ):
-    module_name = _derive_module_name(new_binary_dir)
-    if not callable(call_llm_text):
-        if debug:
-            print("    Preprocess: call_llm_text unavailable for llm_decompile")
-        return _empty_llm_decompile_result()
-
-    if isinstance(symbol_name_list, (list, tuple, set)):
-        symbol_name_text = ", ".join(
-            str(item).strip() for item in symbol_name_list if str(item).strip()
-        )
-    else:
-        symbol_name_text = str(symbol_name_list or "").strip()
-
-    if reference_blocks is None or target_blocks is None:
-        fallback_reference_blocks, fallback_target_blocks = _render_llm_decompile_blocks(
-            [
-                {
-                    "func_name": "Reference",
-                    "disasm_code": disasm_for_reference,
-                    "procedure": procedure_for_reference,
-                }
-            ],
-            [
-                {
-                    "func_name": "Target",
-                    "disasm_code": disasm_code,
-                    "procedure": procedure,
-                }
-            ],
-        )
-        if reference_blocks is None:
-            reference_blocks = fallback_reference_blocks
-        if target_blocks is None:
-            target_blocks = fallback_target_blocks
-
-    if prompt_template is not None:
-        try:
-            prompt = _resolve_llm_decompile_template_value(
-                prompt_template,
-                platform,
-                module_name=module_name,
-            ).format(
-                symbol_name_list=symbol_name_text,
-                disasm_for_reference=str(disasm_for_reference or ""),
-                procedure_for_reference=str(procedure_for_reference or ""),
-                disasm_code=str(disasm_code or ""),
-                procedure=str(procedure or ""),
-                reference_blocks=str(reference_blocks or ""),
-                target_blocks=str(target_blocks or ""),
-                platform=str(platform or "").strip(),
-                module_name=module_name,
-            )
-        except Exception as exc:
-            if debug:
-                print(
-                    f"    Preprocess: failed to format llm_decompile prompt for "
-                    f"{symbol_name_text}: {exc}"
-                )
-            return _empty_llm_decompile_result()
-    else:
-        prompt = (
-            "You are a reverse engineering expert.\n\n"
-            f"Reference functions:\n{reference_blocks}\n\n"
-            f"Target functions:\n{target_blocks}\n\n"
-            f"Please collect all references to \"{symbol_name_text}\" and output YAML."
-        )
-    system_prompt = "You are a reverse engineering expert."
-    if debug:
-        print(
-            f"    Preprocess: calling llm_decompile for {symbol_name_text} "
-            f"with model={str(model).strip()} platform={str(platform or '').strip() or '<empty>'}"
-        )
-        _debug_print_multiline(
-            f"llm_decompile system prompt for {symbol_name_text}",
-            system_prompt,
-            debug=True,
-        )
-        _debug_print_multiline(
-            f"llm_decompile user prompt for {symbol_name_text}",
-            prompt,
-            debug=True,
-        )
-    request_kwargs = {
-        "model": str(model).strip(),
-        "messages": [
-            {"role": "system", "content": system_prompt},
-            {"role": "user", "content": prompt},
-        ],
-        "debug": debug,
-    }
-    if client is not None:
-        request_kwargs["client"] = client
-    try:
-        normalized_temperature = temperature
-        if normalized_temperature is not None and callable(normalize_optional_temperature):
-            normalized_temperature = normalize_optional_temperature(
-                normalized_temperature,
-                "temperature",
-            )
-        if normalized_temperature is not None:
-            request_kwargs["temperature"] = normalized_temperature
-        if effort is not None:
-            request_kwargs["effort"] = effort
-        if api_key is not None:
-            request_kwargs["api_key"] = api_key
-        if base_url is not None:
-            request_kwargs["base_url"] = base_url
-        normalized_fake_as = str(fake_as or "").strip().lower() or None
-        if normalized_fake_as is not None:
-            request_kwargs["fake_as"] = normalized_fake_as
-    except Exception as exc:
-        if debug:
-            print(
-                f"    Preprocess: failed to prepare llm_decompile call for "
-                f"{symbol_name_text}: {exc}"
-            )
-        return _empty_llm_decompile_result()
-
-    max_attempts = _normalize_llm_retry_attempts(max_retries, default=3)
-    delay = _normalize_llm_retry_delay(retry_initial_delay, default=1.0)
-    backoff_factor = _normalize_llm_retry_delay(
-        retry_backoff_factor,
-        default=2.0,
-        minimum=1.0,
+    return await _ida_llm_decompile.call_llm_decompile(
+        client=client,
+        model=model,
+        symbol_name_list=symbol_name_list,
+        disasm_code=disasm_code,
+        procedure=procedure,
+        disasm_for_reference=disasm_for_reference,
+        procedure_for_reference=procedure_for_reference,
+        reference_blocks=reference_blocks,
+        target_blocks=target_blocks,
+        prompt_template=prompt_template,
+        platform=platform,
+        new_binary_dir=new_binary_dir,
+        temperature=temperature,
+        effort=effort,
+        api_key=api_key,
+        base_url=base_url,
+        fake_as=fake_as,
+        max_retries=max_retries,
+        retry_initial_delay=retry_initial_delay,
+        retry_backoff_factor=retry_backoff_factor,
+        retry_max_delay=retry_max_delay,
+        debug=debug,
+        call_llm_text_func=call_llm_text,
+        normalize_temperature_func=normalize_optional_temperature,
     )
-    max_delay = _normalize_llm_retry_delay(retry_max_delay, default=8.0)
-
-    content = None
-    for attempt_index in range(max_attempts):
-        try:
-            content = call_llm_text(**request_kwargs)
-            break
-        except Exception as exc:
-            is_last_attempt = attempt_index >= max_attempts - 1
-            should_retry = _is_transient_llm_error(exc) and not is_last_attempt
-            if not should_retry:
-                if debug:
-                    print(
-                        f"    Preprocess: llm_decompile call failed for "
-                        f"{symbol_name_text}: {exc}"
-                    )
-                return _empty_llm_decompile_result()
-            if debug:
-                print(
-                    f"    Preprocess: llm_decompile transient failure for "
-                    f"{symbol_name_text} on attempt "
-                    f"{attempt_index + 1}/{max_attempts}: {exc}; "
-                    f"retrying in {delay:.2f}s"
-                )
-            if delay > 0:
-                await asyncio.sleep(delay)
-            delay = min(delay * backoff_factor, max_delay)
-    if debug:
-        _debug_print_multiline(
-            f"llm_decompile raw response for {symbol_name_text}",
-            content,
-            debug=True,
-        )
-    parsed_result = parse_llm_decompile_response(content)
-    if debug:
-        _debug_print_json(
-            f"llm_decompile parsed response for {symbol_name_text}",
-            parsed_result,
-            debug=True,
-        )
-    return parsed_result
 
 async def preprocess_vtable_via_mcp(
     session,
